requirements.txt
asgiref==3.8.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2.1
django-allauth==65.8.0
django-pwa==2.0.1
django-tailwind==4.0.1
fonttools==4.58.0
gunicorn==23.0.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.3
mysqlclient==2.2.7
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
